move hive.js inside dhive
