+++ src/screens/community/container/communityContainer.js
@@ -3,8 +3,7 @@ import { withNavigation } from 'react-navigation';
 import get from 'lodash/get';
 import { connect } from 'react-redux';
 
-import { getCommunity, getSubscriptions } from '../../../providers/hive/hive';
-import { subscribeCommunity } from '../../../providers/hive/dhive';
+import { subscribeCommunity, getCommunity, getSubscriptions } from '../../../providers/hive/dhive';
 
 import ROUTES from '../../../constants/routeNames';
