+++ src/screens/searchResult/container/communitiesContainer.js
@@ -5,8 +5,11 @@ import isEmpty from 'lodash/isEmpty';
 
 import ROUTES from '../../../constants/routeNames';
 
-import { getCommunities, getSubscriptions } from '../../../providers/hive/hive';
-import { subscribeCommunity } from '../../../providers/hive/dhive';
+import {
+  getCommunities,
+  getSubscriptions,
+  subscribeCommunity,
+} from '../../../providers/hive/dhive';
 
 const CommunitiesContainer = ({
   children,
